+++ Personas/BasePersona.py
@@ -9,7 +9,6 @@ from Data.Configuration import Configuration
 from Data.EncyclopediaManagement import EncyclopediaManagement
 from Data.FileManagement import FileManagement
 from Data.UserEncyclopediaManagement import UserEncyclopediaManagement
-from Functionality.Organising import Organising
 from Personas.PersonaSpecification.PersonaConstants import SELECT_WORKFLOW_INSTRUCTIONS
 from Utilities.ErrorHandler import ErrorHandler
 from Utilities.Utility import Utility
@@ -33,13 +32,14 @@ class BasePersona:
 
         ErrorHandler.setup_logging()
 
-    def query(self, user_prompt):
+    def query(self, user_prompt, file_references: List[str] = None):
         """
         Handles a user prompt
 
         ToDo adding to the user_encyclopedia needs to be influenced by context category
 
         :param user_prompt: user input prompt
+        :param user_prompt: additional file references paths, format <category_id>/<file_name_with_extension>
         """
         config = Configuration.load_config()
         if Utility.is_valid_prompt(user_prompt):
@@ -47,7 +47,7 @@ class BasePersona:
                 user_encyclopedia_manager = UserEncyclopediaManagement()
                 user_encyclopedia_manager.add_to_encyclopedia([user_prompt])
                 
-            return self.select_workflow(user_prompt)
+            return self.select_workflow(user_prompt, file_references)
         else:
             print("Invalid input. Please ask a clear and valid question.")
 
@@ -85,7 +85,7 @@ class BasePersona:
         for i, (question, response) in enumerate(self.history):
             print(f"{i + 1}: Q: {question}\n    A: {response}")
 
-    def select_workflow(self, initial_message: str):
+    def select_workflow(self, initial_message: str, file_references: List[str] = None):
         """Determine and run the appropriate workflow based on the user's query"""
         executor = AiOrchestrator()
         output = executor.execute_function(
@@ -99,14 +99,21 @@ class BasePersona:
         selected_workflow = output['selection']
 
         logging.info(f"Selected workflow: {selected_workflow}")
-        return self.run_workflow(selected_workflow, initial_message)
+        return self.run_workflow(selected_workflow, initial_message, file_references)
 
-    def run_workflow(self, selected_workflow: str, initial_message: str):
+    def run_workflow(self, selected_workflow: str, initial_message: str, file_references: List[str] = None):
         raise NotImplementedError("This method should be overridden by subclasses")
 
-    def process_question(self, question: str):
+    def process_question(self, question: str, file_references: List[str] = None):
         """Process and store the user's question."""
-        response = self.think([question])
+        file_content = []
+        for file_reference in file_references:
+            content = FileManagement.read_file_full_address(file_reference)
+            logging.info(f"Extracting file content [{file_reference}]: {content}")
+            file_content.append(content)
+
+        input_messages = [question] + file_content
+        response = self.think(input_messages)
         self.history.append((question, response))
         return response